requirements.txt
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
idna==2.8
pkg-resources==0.0.0
python-gitlab==2.0.0
requests==2.22.0
urllib3==1.25.8
